refactored /pages
relocated About, contactUs, Home, Profile, Users to /pages
