+++ github-project/src/pages/contactUs.js
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
-import './App.css';
+import '../App.css';
 
 function ContactUs() {
   const [submitted, setSubmitted] = useState(false);
